+++ src/main/java/org/akhq/utils/avroserde/AvroSchemaDeserializer.java
@@ -1,7 +1,6 @@
-package org.akhq.utils;
+package org.akhq.utils.avroserde;
 
 import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import org.apache.avro.Schema;
