+++ src/main/java/org/akhq/utils/avroserde/AvroSchemaSerializer.java
@@ -1,7 +1,6 @@
-package org.akhq.utils;
+package org.akhq.utils.avroserde;
 
 import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import org.apache.avro.Schema;
