+++ src/main/java/org/akhq/utils/avroserde/BigDecimalFriendlySpecificDatumWriter.java
@@ -1,29 +1,17 @@
-package org.akhq.utils;
+package org.akhq.utils.avroserde;
 
-import org.apache.avro.*;
+import org.apache.avro.Conversion;
+import org.apache.avro.Conversions;
+import org.apache.avro.LogicalType;
+import org.apache.avro.LogicalTypes;
+import org.apache.avro.Schema;
 import org.apache.avro.generic.GenericData;
-import org.apache.avro.generic.GenericRecord;
 import org.apache.avro.io.Encoder;
-import org.apache.avro.io.EncoderFactory;
-import org.apache.avro.io.JsonEncoder;
 import org.apache.avro.specific.SpecificDatumWriter;
 
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.math.BigDecimal;
 import java.nio.ByteBuffer;
-
-public class AvroToJsonSerializer {
-
-    public static String toJson(GenericRecord record) throws IOException {
-        try (ByteArrayOutputStream outputStream = new ByteArrayOutputStream()) {
-            JsonEncoder jsonEncoder = EncoderFactory.get().jsonEncoder(record.getSchema(), outputStream);
-            new BigDecimalFriendlySpecificDatumWriter<GenericRecord>(record.getSchema()).write(record, jsonEncoder);
-            jsonEncoder.flush();
-            return new String(outputStream.toByteArray());
-        }
-    }
-}
 
 class BigDecimalFriendlySpecificDatumWriter<T> extends SpecificDatumWriter<T> {
 
@@ -60,4 +48,3 @@ class BigDecimalFriendlySpecificDatumWriter<T> extends SpecificDatumWriter<T> {
         }
     }
 }
-
