+++ src/test/java/org/akhq/modules/AvroSchemaSerializerTest.java
@@ -2,6 +2,7 @@ package org.akhq.modules;
 
 import io.confluent.kafka.schemaregistry.client.SchemaRegistryClient;
 import io.confluent.kafka.schemaregistry.client.rest.exceptions.RestClientException;
+import org.akhq.utils.avroserde.AvroSerializer;
 import org.apache.avro.AvroTypeException;
 import org.apache.avro.SchemaBuilder;
 import org.junit.jupiter.api.BeforeEach;
